chore: rework visual testing for select
