+++ src/components/inputs/select-input/select-input.visualroute.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Route, Switch } from 'react-router-dom';
 import { SelectInput } from 'ui-kit';
 import { Suite, Spec } from '../../../../test/percy';
 
@@ -7,11 +8,16 @@ const options = [
   { value: 'two', label: 'Two' },
 ];
 
+const optionsWithGroups = [
+  { label: 'one', options: [{ value: 'one', label: 'One' }] },
+  { options: [{ value: 'two', label: 'Two' }] },
+];
+
 const value = 'one';
 
 export const routePath = '/select-input';
 
-export const component = () => (
+const DefaultRoute = () => (
   <Suite>
     <Spec label="minimal">
       <SelectInput
@@ -68,4 +74,62 @@ export const component = () => (
       />
     </Spec>
   </Suite>
+);
+
+const OpenRoute = () => (
+  <Suite>
+    <Spec label="option without option groups">
+      <SelectInput
+        id="select-input"
+        value={value}
+        onChange={() => {}}
+        options={options}
+        horizontalConstraint="m"
+      />
+    </Spec>
+  </Suite>
+);
+
+const OpenRouteWithOptionGroups = () => (
+  <Suite>
+    <Spec label="option group with no divider">
+      <SelectInput
+        id="select-input"
+        value={value}
+        onChange={() => {}}
+        options={optionsWithGroups}
+        horizontalConstraint="m"
+      />
+    </Spec>
+  </Suite>
+);
+
+const OpenRouteWithOptionGroupsAndDivider = () => (
+  <Suite>
+    <Spec label="option group with divider">
+      <SelectInput
+        id="select-input"
+        value={value}
+        onChange={() => {}}
+        options={optionsWithGroups}
+        horizontalConstraint="m"
+        showOptionGroupDivider={true}
+      />
+    </Spec>
+  </Suite>
+);
+
+export const component = () => (
+  <Switch>
+    <Route path={`${routePath}/open`} component={OpenRoute} />
+    <Route
+      path={`${routePath}/open-with-option-groups`}
+      component={OpenRouteWithOptionGroups}
+    />
+    <Route
+      path={`${routePath}/open-with-option-groups-and-divider`}
+      component={OpenRouteWithOptionGroupsAndDivider}
+    />
+    <Route path={routePath} component={DefaultRoute} />
+  </Switch>
 );
